cohorts/2023/week_2_workflow_orchestration/week_2_code/requirements.txt
pandas==1.5.2
prefect-gcp[cloud_storage]==0.2.4
psycopg2-binary==2.9.5
prefect-sqlalchemy==0.2.2
pyarrow==10.0.1
pandas-gbq==0.18.1